+++ NEWS.md
@@ -21,6 +21,8 @@ New features(Analysis):
   (these checks can also be enabled with the config setting `redundant_condition_detection`)
 + New issue types: `PhanImpossibleTypeComparison` (when `--redundant-condition-detection` is enabled) (#1807)
   (e.g. warns about `$x = new stdClass(); assert($x !== null)`)
++ New issue types: `PhanCoalescingAlwaysNull`, `PhanCoalescingNeverNull` (when `--redundant-condition-detection` is enabled)
+  (e.g. warns about `(null ?? 'other')`, `($a >= $b) ?? 'default'`)
 + Make Phan more accurately check if a loop may be executed 0 times.
 + Make Phan more accurately check if a loop may be executed 0 times.
 + Support the type `callable-object` in phpdoc and infer it from checks such as `is_callable($var) && is_object($var)` (#1336)